+++ game/static/game/js/level_editor.js
@@ -983,8 +983,8 @@ ocargo.LevelEditor = function(levelId) {
         var windowWidth = $(window).width();
         var windowHeight = $(window).height();
 
-        var paperRightEdge = PAPER_WIDTH + $('#tools').width();
-        var paperBottomEdge = PAPER_HEIGHT;
+        var paperRightEdge = EXTENDED_PAPER_WIDTH + $('#tools').width();
+        var paperBottomEdge = EXTENDED_PAPER_HEIGHT;
 
         var bottom = 50;
         if(windowHeight > paperBottomEdge) {
@@ -1465,14 +1465,14 @@ ocargo.LevelEditor = function(levelId) {
             // Stop it being dragged off the edge of the page
             if (paperX < 0) {
                 paperX = 0;
-            } else if (paperX + imageWidth > paperWidth) {
-                paperX = paperWidth - imageWidth;
+            } else if (paperX + imageWidth > EXTENDED_PAPER_WIDTH) {
+                paperX = EXTENDED_PAPER_WIDTH - imageWidth;
             }
 
             if (paperY < 0) {
                 paperY = 0;
-            } else if (paperY + imageHeight >  paperHeight) {
-                paperY = paperHeight - imageHeight;
+            } else if (paperY + imageHeight >  EXTENDED_PAPER_HEIGHT) {
+                paperY = EXTENDED_PAPER_HEIGHT - imageHeight;
             }
 
             image.transform('t' + paperX + ',' + paperY);
@@ -1495,8 +1495,17 @@ ocargo.LevelEditor = function(levelId) {
             originX = paperX;
             originY = paperY;
 
-            if(trashcanOpen) {
+            if (trashcanOpen) {
                 decor.destroy();
+            } else {
+                if (paperWidth < paperX + imageWidth) {
+                    originX = paperWidth - imageWidth;
+                }
+                if (paperHeight < paperY + imageHeight) {
+                    originY = paperHeight - imageHeight;
+                }
+
+                image.transform('t' + originX + ',' + originY);
             }
 
             closeTrashcan();
@@ -1546,13 +1555,13 @@ ocargo.LevelEditor = function(levelId) {
             // Stop it being dragged off the edge of the page
             if (paperX < 0) {
                 paperX = 0;
-            } else if (paperX + imageWidth > paperWidth) {
-                paperX = paperWidth - imageWidth;
+            } else if (paperX + imageWidth > EXTENDED_PAPER_WIDTH) {
+                paperX = EXTENDED_PAPER_WIDTH - imageWidth;
             }
             if (paperY < 0) {
                 paperY =  0;
-            } else if (paperY + imageHeight >  paperHeight) {
-                paperY = paperHeight - imageHeight;
+            } else if (paperY + imageHeight >  EXTENDED_PAPER_HEIGHT) {
+                paperY = EXTENDED_PAPER_HEIGHT - imageHeight;
             }
 
             // And perform the updatee
@@ -1652,9 +1661,9 @@ ocargo.LevelEditor = function(levelId) {
                 }
             }
 
-            if(trashcanOpen) {
+            if (trashcanOpen) {
                 cow.destroy();
-            } else if(isValidPlacement(controlledCoord)) {
+            } else if (isValidPlacement(controlledCoord)) {
                 // Add back to the list of cows if on valid nodes
                 var controlledNode = ocargo.Node.findNodeByCoordinate(controlledCoord, nodes);
                 cow.controlledNode = controlledNode;
@@ -1663,7 +1672,20 @@ ocargo.LevelEditor = function(levelId) {
             } else {
                 cow.controlledNode = null;
                 cow.valid = false;
-            }
+
+                var cowX = paperX;
+                var cowY = paperY;
+
+                if (paperWidth < paperX + imageWidth) {
+                    cowX = paperWidth - imageWidth
+                }
+
+                if (paperHeight < paperY + imageHeight) {
+                    cowY = paperHeight - imageHeight
+                }
+
+                image.transform('t' + cowX + ',' + cowY);
+            } 
             adjustCowGroupMinMaxFields(cow);
 
             image.attr({'cursor':'pointer'});
@@ -1753,16 +1775,16 @@ ocargo.LevelEditor = function(levelId) {
             // Stop it being dragged off the edge of the page
             if (paperX < 0) {
                 paperX = 0;
-            } else if (paperX + imageWidth > paperWidth) {
-                paperX = paperWidth - imageWidth;
+            } else if (paperX + imageWidth > EXTENDED_PAPER_WIDTH) {
+                paperX = EXTENDED_PAPER_WIDTH - imageWidth;
             }
             if (paperY < 0) {
                 paperY =  0;
-            } else if (paperY + imageHeight >  paperHeight) {
-                paperY = paperHeight - imageHeight;
+            } else if (paperY + imageHeight >  EXTENDED_PAPER_HEIGHT) {
+                paperY = EXTENDED_PAPER_HEIGHT - imageHeight;
             }
 
-            // And perform the updatee
+            // And perform the update
             image.transform('t' + paperX + ',' + paperY + 'r' + rotation + 's' + scaling);
 
             // Unmark the squares the light previously occupied
@@ -1912,6 +1934,18 @@ ocargo.LevelEditor = function(levelId) {
                 trafficLight.valid = true;
 
                 drawing.setTrafficLightImagePosition(sourceCoord, controlledCoord, image);
+            } else {
+                var trafficLightX = paperX;
+                var trafficLightY = paperY;
+
+                if (paperWidth < paperX + imageWidth) {
+                    trafficLightX = paperWidth - imageWidth
+                    image.transform('t' + trafficLightX + ',' + trafficLightY + 'r' + rotation + 's' + scaling);
+                }
+                if (paperHeight < paperY + imageHeight) {
+                    trafficLightY = paperHeight - imageHeight
+                    image.transform('t' + trafficLightX + ',' + trafficLightY + 'r' + rotation + 's' + scaling);
+                }
             }
 
             image.attr({'cursor':'pointer'});