+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 <!--next-version-placeholder-->
 
+## v5.17.1 (2024-05-13)
+
+### Fix
+
+* Fix wonky level editor ([#1609](https://github.com/ocadotechnology/rapid-router/issues/1609)) ([`426f801`](https://github.com/ocadotechnology/rapid-router/commit/426f801fa309615e8217a4e213c72c446540ff55))
+
+## v5.17.0 (2024-05-09)
+
+### Feature
+
+* Add language dropdown ([#1606](https://github.com/ocadotechnology/rapid-router/issues/1606)) ([`71d5d88`](https://github.com/ocadotechnology/rapid-router/commit/71d5d88a4906c4bf6efc5a631d6f1a62dd93a4cb))
+
 ## v5.16.25 (2024-05-07)
 
 ### Fix